pyproject.toml
[project]
name = "healthpadi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
]
